+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Add Windows backend #99
 - Add pager support on Windows
+- Add Intel RSD Network card information #81
 
 ### Fixed